[PRE-RELEASE PATCH 3]CityLife Core 新功能添加, CL Mood Debug 新进展.
本次更新改动:
  > CityLife Core
    - 新增枚举类 ChatRequestDataType 以用于验证请求的数据类型.
    - 数据类 InventoryItem 新增 hasChatRequestAction 方法判断是否含有聊天栏数据请求动作以分类执行动作处理.
    - 监听器 PlayerInteractListener 修复了区域选择器选择时会把方块打破的问题(创造模式), 新增 方块黑名单 功能的实现代码.
    - 管理器 ChatRequestManager 注册方法新增对应请求数据类型参数.
    - 工具类 InventoryUtil 新增 rebuildInventory 方法以重构 Gui.
    - 工具类 ItemStackUtil 优化算法以适配 CL Friends 需求及解决潜在的 Bug.
    - 配置文件 Config.yml 新增新功能的配置选项.
    - 配置文件 Language.yml & Config.yml 修改了过时的插件前缀.

  > CityLife Friends
    - 重写 Gui 配置文件管理器 中的大多数方法.
    - 以及其他 Gui 管理器 的核心方法改动等.

  > CityLife Mood
    - 修复部分子命令在发送提示信息时的报错.
    - 修复读取 Area 时的报错.
    - 修复读取随机操作时的报错.
    - 修复了一个可能导致消耗品失效的问题.
    - 配置文件 Language.yml 新增了部分语言信息.

  > SoulRing
    - 更改了数据读取方式, 全部使用 int 型整数.
    - 适配上述更改重写了部分属性读取方法.
    - 新属性: HEALTH(生命力), 可以提升生命值上限.[UNFINISHED]

+++ SoulRing/src/main/java/org/serverct/sir/soulring/listener/OnPlayerAttack.java
@@ -4,7 +4,6 @@ import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.EntityDamageByEntityEvent;
-import org.bukkit.inventory.ItemStack;
 import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
 import org.serverct.sir.soulring.Attributes;
@@ -17,8 +16,8 @@ import java.util.Random;
 
 public class OnPlayerAttack implements Listener {
 
-    private Map<Attributes, Double> damagerAttributesMap = new HashMap<>();
-    private Map<Attributes, Double> targetAttributesMap = new HashMap<>();
+    private Map<Attributes, Integer> damagerAttributesMap = new HashMap<>();
+    private Map<Attributes, Integer> targetAttributesMap = new HashMap<>();
 
     Random r = new Random();
 
@@ -40,7 +39,7 @@ public class OnPlayerAttack implements Listener {
             if(damagerAttributesMap.keySet().contains(Attributes.VAMPIRE_RATE) && damagerAttributesMap.keySet().contains(Attributes.VAMPIRE_PERCENT)) {
                 if(r.nextInt(100) <= damagerAttributesMap.get(Attributes.VAMPIRE_RATE)) {
                     // Bukkit.getLogger().info(ChatColor.translateAlternateColorCodes('&', " &d> &7生命吸取触发."));
-                    value = Math.ceil(evt.getDamage() * ((damagerAttributesMap.get(Attributes.VAMPIRE_PERCENT) / 100) + 1));
+                    value = Math.ceil(evt.getDamage() * ((double) (damagerAttributesMap.get(Attributes.VAMPIRE_PERCENT) / 100) + 1));
                     color = AttributeManager.getInstance().getColor(Attributes.VAMPIRE_RATE).toString();
                     // Bukkit.getLogger().info(ChatColor.translateAlternateColorCodes('&', " &d> &7回复血量 " + String.valueOf(heal) + " ."));
                     if(damager.getHealth() + value <= damager.getMaxHealth()) {
